Added is_valid_config flag to rejig_state.
This is to prevent clients from accessing stale rejig configs after the
config id was updated. Since process_update_command is itself
asyncronous, we can't wait on it to finish executing.

Also fixed a bug that prevented copying of the rejig config id while
storing items that change slab classes on prepend, or append.